# Conflicts:
#   CMakeLists.txt

+++ Kernel/LHCbKernel/Kernel/RichSide.h
@@ -16,7 +16,7 @@
 #include <string>
 #include <vector>
 #include <iostream>
-#include <stdint.h>
+#include <cstdint>
 #include <array>
 
 // Kernel
@@ -26,10 +26,10 @@ namespace Rich
 {
 
   /// Total number of PD panels
-  static const unsigned int NTotalPDPanels   = 4;
+  static const uint8_t NTotalPDPanels   = 4;
 
   /// Number of PD panels per RICH detector
-  static const unsigned int NPDPanelsPerRICH = 2;
+  static const uint8_t NPDPanelsPerRICH = 2;
 
   /** @enum Rich::Side
    *
